+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(logitnorm)
+
+test_check("logitnorm")