+++ tests/testthat/test_modeLogitnorm.R
@@ -1,12 +1,4 @@
-#TODO
-
-.setUp <-function () {
-}
-
-.tearDown <- function () {
-}
-
-test.rightMode <- function(){
+test_that("rightMode", {
 	theta0 <- c(mu=1.5, sigma=0.8)
 	#plot the true and the rediscovered distributions
 	xGrid = seq(0,1, length.out=81)[-c(1,81)]
@@ -20,11 +12,11 @@ test.rightMode <- function(){
 	#check by monte carlo integration
 	#z <- rlogitnorm(1e6, mu=theta0[1], sigma=theta0[2]);	var(z)
 	#dz <- density(z)
-	#checkEqualsNumeric( dz$x[which.max(dz$y)], mle, tolerance=5e-2)
-	checkEqualsNumeric( 0.88, mle, tolerance=1e-2)
-}
+	#expect_equal( dz$x[which.max(dz$y)], mle, tolerance=5e-2)
+	expect_equal( 0.88, mle, tolerance=1e-2)
+})
 
-test.leftMode <- function(){
+test_that("leftMode", {
 	theta0 <- c(mu=-1.5, sigma=0.8)
 	#plot the true and the rediscovered distributions
 	xGrid = seq(0,1, length.out=81)[-c(1,81)]
@@ -39,7 +31,7 @@ test.leftMode <- function(){
 	# deprecated: did not run on Windows
 	#z <- rlogitnorm(1e6, mu=theta0[1], sigma=theta0[2]);	var(z)
 	#dz <- density(z)
-	#checkEqualsNumeric( dz$x[which.max(dz$y)], mle, tolerance=5e-2)
-	checkEqualsNumeric( 0.12, mle, tolerance=1e-2)		# regression 0.12 calculated previously
-}
+	#expect_equal( dz$x[which.max(dz$y)], mle, tolerance=5e-2)
+	expect_equal( 0.12, mle, tolerance=1e-2)		# regression 0.12 calculated previously
+})